cost-metrics/requirements.txt
boto3
requests
diskcache>=5.6.0
ec2-metadata>=2.11.0
pyyaml>=6.0.0